+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to the "ev3dev-browser" extension will be documented in this
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## Unreleased
+### Fixed
+- Fix debugger hanging when ev3dev Device Browser view is collapsed
+
 ## 1.0.2 - 2019-03-11
 ### Fixed
 - Files are not downloaded when using global launch configuration